Refactor general statistics
